+++ anchor/keygen/Cargo.toml
@@ -16,4 +16,5 @@ serde_json = { workspace = true }
 sha2 = { workspace = true }
 thiserror = { workspace = true }
 tracing = { workspace = true }
+tracing-subscriber = { workspace = true }
 zeroize = { workspace = true }
