+++ anchor/keysplit/Cargo.toml
@@ -24,4 +24,5 @@ sha2 = { workspace = true }
 ssv_types = { workspace = true }
 tokio = { workspace = true }
 tracing = { workspace = true }
+tracing-subscriber = { workspace = true }
 types = { workspace = true }
